fixes for python3

+++ src/DataGenerator.py
@@ -29,7 +29,7 @@ class DataGenerator(keras.utils.Sequence):
 
     def __len__(self):
         'Denotes the number of batches per epoch'
-        return int(np.floor(self.num_of_traj*(self.number_of_frames-1) / self.batch_size))/2
+        return int(int(np.floor(self.num_of_traj*(self.number_of_frames-1) / self.batch_size))/2)
 
     def __getitem__(self, index):
         'Generate one batch of data'
@@ -132,7 +132,7 @@ class RelationDataGeneratorMany(keras.utils.Sequence):
 
     def __len__(self):
         'Denotes the number of batches per epoch'
-        return int(np.floor(self.num_of_traj*(self.number_of_frames-self.timestep_diff) / self.batch_size))/2
+        return int(int(np.floor(self.num_of_traj*(self.number_of_frames-self.timestep_diff) / self.batch_size))/2)
 
     def __getitem__(self, index):
         'Generate one batch of data'
